restructure, gen a lockfile and push
